refactor: migrate ArmySplit harness and low-risk e2e mirrors (#3847)
Route ArmySplitTestHarness through buildAppShell, consolidate the sole
low-risk e2e mirror into a barrel smoke test, and document stay-split
partN families plus approved app test support harnesses in repo-lint.
